Move Circle/Bullet to a models file.
