+++ src/scene/update.js
@@ -1,7 +1,8 @@
 import Phaser from 'phaser'
 import {CONFIG} from '../config'
 import {WORLD} from '../world'
-import {Bullet, applyVector, ephemeraInBounds, throtLog} from '../util'
+import {Bullet} from '../models'
+import {applyVector, ephemeraInBounds, throtLog} from '../util'
 import {fireUp, fireDown, fireLeft, fireRight} from '../ephemera/bullets'
 import {addPlusOne} from '../ephemera/powerups'
 import {updateEphemeraPost, updateEphemeraPre, updatePre} from '../features'
